✨ Adds navigation logic (router)
