+++ src/App.tsx
@@ -1,12 +1,16 @@
 import "./App.css";
-import Login from "./pages/Home/Login/Login";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import Login from "./pages/Login/Login";
+import Profile from "./pages/Profile/Profile";
 
 function App() {
   return (
-    <div className="main-container">
-      <h2>Sign in to ProfileHub</h2>
-      <Login />
-    </div>
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Login />}></Route>
+        <Route path="/profile" element={<Profile />}></Route>
+      </Routes>
+    </BrowserRouter>
   );
 }
